signing_and_verification/requirements.txt
fire==0.4.0
PyNaCl==1.5.0